Add main-nav custom element
